pyproject.toml
# pyproject.toml

[tool.poetry]
name = "scanless"
version = "0.1.0"
description = ""

[tool.poetry.dependencies]
python = "^3.9"
python-sane = "2.9.1"
pytesseract = "0.3.13"
uvicorn = "0.35.0"
fastapi = "0.116.0"
Jinja2 = "3.1.6"
pillow = "^11.3.0"
pydantic = "^2.11.7"
pypdf2 = "^3.0.1"

[poetry.group.dev.dependencies]
